Optimize feeds

+++ app/decorators/item_decorator.rb
@@ -116,20 +116,20 @@ class ItemDecorator < ApplicationDecorator
 
     if h.current_user.present?
       #check for bookmarks
-      if h.current_user.bookmarks.map(&:item).include?(item)
-        bookmark = h.current_user.bookmarks.detect { |bookmark| bookmark.item == item }
+      if item.bookmark_users.include?(h.current_user)
+        bookmark = item.bookmarks.detect { |bookmark| bookmark.user_id == current_user.id }
         links << link_to('Remove Bookmark', bookmark_path(bookmark), :method => :delete)
       else
         links << link_to('Bookmark', bookmarks_path(:item_id => item.id), :method => :post)
       end
     end
 
-    if h.current_user
+    if h.current_user && item.type == 'Event'
       #check for rsvps
-      if h.current_user.rsvps.map(&:item).include?(item)
-        rsvp = Rsvp.find_by_user_id_and_item_id(h.current_user.id, item.id)
+      if item.rsvp_users.include?(h.current_user)
+        rsvp = item.rsvps.where(:user_id => h.current_user.id)
         links << link_to('Cancel RSVP', rsvp_path(rsvp), :method => :delete)
-      elsif item.type == 'Event'
+      else
         links << link_to('RSVP', rsvps_path(:item_id => item.id), :method => :post)
       end
     end
